@@ -1,4 +1,3 @@
-# built-in
 from typing import Any, Callable, Generic, Optional, Type, TypeVar, overload
 
 
@@ -6,10 +5,8 @@ T = TypeVar('T')
 
 
 class cached_property(Generic[T]):  # noqa: N801
-    func: Callable[[Any], T]
-
     def __init__(self, func: Callable[[Any], T]) -> None:
-        self.func = func  # type: ignore
+        self.func = func
 
     @overload
     def __get__(self, instance: None, owner: Optional[Type[Any]] = ...) -> 'cached_property[T]':
@@ -20,5 +17,6 @@ class cached_property(Generic[T]):  # noqa: N801
         pass
 
     def __get__(self, obj, cls):
-        value = obj.__dict__[self.func.__name__] = self.func(obj)
+        value = self.func(obj)
+        obj.__dict__[self.func.__name__] = value
         return value
