@@ -1,14 +1,13 @@
-# built-in
 from argparse import ArgumentParser
 from pathlib import Path
-from typing import List, Sequence
+from typing import List
 
-# app
 from ..linter import StubsManager, generate_stub
+from ._base import Command
 from ._common import get_paths
 
 
-def stub_command(argv: Sequence[str]) -> int:
+class StubCommand(Command):
     """Generate stub files for the given Python files.
 
     ```bash
@@ -25,20 +24,22 @@ def stub_command(argv: Sequence[str]) -> int:
 
     [stubs]: https://deal.readthedocs.io/details/stubs.html
     """
-    parser = ArgumentParser(prog='python3 -m deal stub')
-    parser.add_argument('--iterations', type=int, default=1)
-    parser.add_argument('paths', nargs='+')
-    args = parser.parse_args(argv)
 
-    paths: List[Path] = []
-    for arg in args.paths:
-        for path in get_paths(Path(arg)):
-            paths.append(path)
+    @staticmethod
+    def init_parser(parser: ArgumentParser) -> None:
+        parser.add_argument('--iterations', type=int, default=1)
+        parser.add_argument('paths', nargs='+')
 
-    roots = list(StubsManager.default_paths) + list(set(paths))
-    stubs = StubsManager(paths=roots)
+    def __call__(self, args) -> int:
+        paths: List[Path] = []
+        for arg in args.paths:
+            for path in get_paths(Path(arg)):
+                paths.append(path)
 
-    for _ in range(args.iterations):
-        for path in paths:
-            generate_stub(path=path, stubs=stubs)
-    return 0
+        roots = list(StubsManager.default_paths) + list(set(paths))
+        stubs = StubsManager(paths=roots)
+
+        for _ in range(args.iterations):
+            for path in paths:
+                generate_stub(path=path, stubs=stubs)
+        return 0
