@@ -1,7 +1,6 @@
-# built-in
-from typing import Callable, Type, TypeVar
+from typing import Callable, Iterator, Type, TypeVar
 
-# app
+from .._cached_property import cached_property
 from .._exceptions import ReasonContractError
 from .._types import ExceptionType
 from .base import Base
@@ -25,6 +24,12 @@ class Reason(Base[_CallableType]):
             exception=exception,
         )
 
+    @cached_property
+    def exception_type(self) -> Type[Exception]:
+        if isinstance(self.exception, Exception):
+            return type(self.exception)
+        return self.exception
+
     def patched_function(self, *args, **kwargs):
         """
         Step 3. Wrapped function calling.
@@ -34,7 +39,7 @@ class Reason(Base[_CallableType]):
         except self.event as origin:
             try:
                 self.validate(*args, **kwargs)
-            except self.exception:
+            except self.exception_type:
                 raise self.exception from origin
             raise
 
@@ -47,11 +52,11 @@ class Reason(Base[_CallableType]):
         except self.event as origin:
             try:
                 self.validate(*args, **kwargs)
-            except self.exception:
+            except self.exception_type:
                 raise self.exception from origin
             raise
 
-    def patched_generator(self, *args, **kwargs):
+    def patched_generator(self, *args, **kwargs) -> Iterator:
         """
         Step 3. Wrapped function calling.
         """
@@ -60,6 +65,6 @@ class Reason(Base[_CallableType]):
         except self.event as origin:
             try:
                 self.validate(*args, **kwargs)
-            except self.exception:
+            except self.exception_type:
                 raise self.exception from origin
             raise
