@@ -1,13 +1,10 @@
-# built-in
 import ast
 import sys
 from types import ModuleType
 from typing import Any, Callable, List, Optional
 
-# external
-from _frozen_importlib_external import PathFinder
+from _frozen_importlib_external import PathFinder  # pyright: reportMissingImports=false
 
-# app
 from . import _aliases
 from ._state import state
 
@@ -68,7 +65,7 @@ class DealLoader:
         wrapped(module)
 
     @staticmethod
-    def _get_contracts(tree: ast.Module) -> List[ast.AST]:
+    def _get_contracts(tree: ast.Module) -> List[ast.expr]:
         for node in tree.body:  # type: Any
             if type(node) is not ast.Expr:
                 continue
@@ -95,7 +92,7 @@ class DealLoader:
 
         if not isinstance(node, ast.Attribute):
             return None
-        if node.value.id != 'deal':  # type: ignore
+        if node.value.id != 'deal':
             return None
         contract = getattr(_aliases, node.attr, None)
         if contract is None:
@@ -155,7 +152,7 @@ def activate() -> bool:
     if DealFinder in sys.meta_path:
         return False
     index = sys.meta_path.index(PathFinder)
-    sys.meta_path[index] = DealFinder  # type: ignore
+    sys.meta_path[index] = DealFinder
     return True
 
 
@@ -164,6 +161,6 @@ def deactivate() -> bool:
     """
     if DealFinder not in sys.meta_path:
         return False
-    index = sys.meta_path.index(DealFinder)  # type: ignore
+    index = sys.meta_path.index(DealFinder)
     sys.meta_path[index] = PathFinder
     return True
