@@ -1,15 +1,12 @@
-# built-in
 import typing
 from functools import update_wrapper
 from inspect import signature
 
-# external
 import hypothesis
 import hypothesis.strategies
 import typeguard
 from hypothesis.internal.reflection import proxies
 
-# app
 from ._cached_property import cached_property
 from ._decorators import Pre, Raises
 from ._types import ArgsKwargsType
@@ -52,7 +49,7 @@ class TestCase(typing.NamedTuple):
         try:
             result = self.func(*self.args, **self.kwargs)
         except self.exceptions:
-            return typing.NoReturn  # type: ignore
+            return typing.NoReturn
         self._check_result(result)
         return result
 
@@ -325,7 +322,7 @@ class cases:  # noqa: N
             return None
         if callable(target):
             return self._wrap(target)
-        return self._run.hypothesis.fuzz_one_input(target)
+        return self._run.hypothesis.fuzz_one_input(target)  # type: ignore[attr-defined]
 
     # a hack to make the test discoverable by pytest
     @property
