@@ -1,4 +1,3 @@
-# built-in
 import ast
 import inspect
 import sys
@@ -37,7 +36,7 @@ def trace(func, **kwargs) -> TraceResult:
 
     old_trace = sys.gettrace()
     try:
-        func_result: Any = t.runfunc(func, **kwargs)  # type: ignore
+        func_result: Any = t.runfunc(func, **kwargs)
     finally:
         # restore previous tracer
         sys.settrace(old_trace)     # pragma: no cover
