@@ -1,12 +1,9 @@
-# built-in
 import ast
 import typing
 from pathlib import Path
 
-# external
 from astroid import AstroidSyntaxError
 
-# app
 from ._error import Error
 from ._func import Func
 from ._rules import Required, rules
@@ -29,14 +26,13 @@ class Checker:
 
     @property
     def version(self):
-        # project
         import deal
 
         return deal.__version__
 
     def run(self) -> typing.Iterator[tuple]:
         for error in self.get_errors():
-            yield tuple(error) + (type(self),)  # type: ignore
+            yield tuple(error) + (type(self),)
 
     def get_funcs(self) -> typing.List['Func']:
         if self._filename == 'stdin':
