@@ -1,9 +1,7 @@
-# built-in
 import json
 from pathlib import Path
 
-# project
-from deal._cli._stub import stub_command
+from deal._cli import main
 
 
 TEXT = """
@@ -17,7 +15,7 @@ def f(x):
 def test_stub_command(tmp_path: Path):
     path = (tmp_path / 'example.py')
     path.write_text(TEXT)
-    result = stub_command(['--iterations', '1', str(path)])
+    result = main(['stub', '--iterations', '1', str(path)])
     assert result == 0
     content = (tmp_path / 'example.json').read_text()
     assert json.loads(content) == {'f': {'raises': ['PatheticError']}}
