@@ -1,10 +1,7 @@
-# external
 import pytest
 
-# project
 import deal
 
-# app
 from .helpers import run_sync
 
 
@@ -31,6 +28,14 @@ def test_reason_just_works():
 def test_reason_exception(value, exc):
     with pytest.raises(exc):
         func(value)
+
+
+def test_reason_custom_exc():
+    @deal.reason(KeyError, lambda: False, exception=ValueError('hello'))
+    def func():
+        raise KeyError
+    with pytest.raises(ValueError):
+        func()
 
 
 @pytest.mark.parametrize('value, exc', [
